asasas

+++ Project/GameProject/Game/Game.h
@@ -4,7 +4,7 @@
 class Game : public Base {
 private:
 	CImage m_BackGround;
-	int m_cnt;
+	float m_cnt;
 	int m_score;
 public:
 	Game();
